Fix: signal handling (#305)
* fix signal handling

* Update feedbacks

* Fix feedbacks
