+++ waffle_hub/__init__.py
@@ -1,7 +1,6 @@
 __version__ = "0.3.1a1"
 
 import enum
-import signal
 from collections import OrderedDict
 
 
@@ -159,16 +158,3 @@ for key in list(EXPORT_MAP.keys()):
 for key in list(BACKEND_MAP.keys()):
     BACKEND_MAP[str(key).lower()] = BACKEND_MAP[key]
     BACKEND_MAP[str(key).upper()] = BACKEND_MAP[key]
-
-
-# except handler for SIGINT, SIGTERM, SIGCHILD
-def sigint_handler(signum, frame):
-    raise KeyboardInterrupt
-
-
-def sigterm_handler(signum, frame):
-    raise SystemExit
-
-
-signal.signal(signal.SIGINT, sigint_handler)
-signal.signal(signal.SIGTERM, sigterm_handler)
